Ajout des ameliorations de fusil

+++ classes/game-objects/Monstre.cs
@@ -13,7 +13,7 @@ public class Monstre : AbstractGameObject
     private float _vitesse;
     private float _vitesseRot;
     private float _rayon;
-    private int hp;
+    private float hp;
 
     public Monstre(Vector2[] forme, Vector2 position, EcranJeu ecranJeu, float rayon)
         : base(forme, position, 1)
@@ -75,7 +75,7 @@ public class Monstre : AbstractGameObject
     /// </summary>
     /// <param name="degat">Nombre de degat subit</param>
     /// <returns>true si le monstre est mort</returns>
-    public bool RecevoirDegat(int degat){
+    public bool RecevoirDegat(float degat){
         this.hp -= degat;
         if(hp <= 0){
             Mourrir();
